requirements.txt
biopython
requests